service/requirements.txt
numpy==1.17.4
pandas==0.24.2
boto3==1.12.39
flask==1.1.1
flask-cors==3.0.8
Pillow==6.2.0
requests==2.22.0
jinja2==2.10.3
awscli==1.16.14